Pre reconfigure of BasePage to StatefulWidget.
Currently the page contents gets built at the same time as ButtonArray. But this means that ButtonArray Rect data is not available during the initial ListView build for SettingsPageContents.

Idea is then to have a contents variable, set to null for initial BasePage build, and then to immediately set state with non-null page contents.