Implement fully functional SettingsPageListTileBorder.
SettingsPageListTile is set, for diagnostic purposes, so that its radius is tiny in order to make the corners obvious. The corners are the points which move around ButtonArray and generate the shrink-expand effect.

SettingsPageListTileBorder has been tested for all alignments with ButtonArray vertical. Need to test for a horizontal ButtonArray.

The circularly curved section was tested by plotting (deltaX, y) data in LibreOffice.