Use getDeltaXFromLowerLocalConstructionRect in the calculation getDeltaXFromUpperLocalConstructionRect.
The existing methods, getDeltaXFromLowerLocalConstructionRect and getDeltaXFromUpperLocalConstructionRect share a a substantial amount of code. Specifically, both methods calculate a directionless deltaX which is then interpreted and applied correctly in _getPath.

In this commit the y-translation is extracted from getDeltaXFromLowerLocalConstructionRect and getDeltaXFromUpperLocalConstructionRect, placed in _getPath, so that getDeltaXFromUpperLocalConstructionRect can be made to call getDeltaXFromLowerLocalConstructionRect.