Rectify offset associated with ListViewBuilderSettingsPageListTile.
The original method for obtaining rect, representing the extent of the Container which this widget is based upon assumed that the origin was co-located with the ListView top left corner.

In actual fact the origin used to construct the container was the top left corner of the screen, thus iplying an offset.

Thise bug was corrected by using an instance of DataNotifer to pass down the Rect data associated with the Stack in BasePage.