Rename getDeltaXFromPosition as getDeltaXFromY and add a radius input.
With the height of lowerConstrRect or upperConstrRect set greater than 1, in units of guestRect!.shortestSide, then the radius associated with the curve was determined by the size of construction Rect and not its proper value of guestRect!.shortestSide / 2. This bug was visible by virtue of the infinite horizontal acceleration given to the tiles for any vertical size of construction Rect.