Reimplement SettingsPageListTile with a shrinkable test Text widget using a Row-Expanded-Stack-Align construct instead of the original Stack-Align-Row-Expanded approach.
In both cases the fade effect is implemented by two instances of Align at the levels of the widget tree indicated by the Align in the above example constructs.

The difference between the two approaches is that there should be less to render in this new approach due to the Align being at the top of the tree.