Implement and test NotificationDataStore.
In main an instance of NotificationDataStore catches notifications of type DataNotification bubbling up the widget tree and prints a message.

Note there are two other widgets that listen for DataNotification. If the callback associated with any of these returns true then DataNotification is prevented from proceeding up the widget tree and NotificationDataStore will not register.