Working example of base settings page functionality.
Convert _KarKam to a SatefulWidget and use setState to trigger a rebuild of the entire app.

When pulling down a copy of an InheritedWidget and changing its value, this changes the value of the InheritedWidget further up the widget tree. This implemetation of base settings page functionality takes a copy of the pulled down version, alters the copy and transmits it back via a notification. In _KarKam a comparison is made and the original value updated with setState.