Add new file settings.dart.
The StatefulWidget, Settings, in settings.dart constitutes a reconfiguration of AppModel. Currently Settings is a functionless wrapper widget for the Material widget in _KarKamState, but the intention is to migrate all elements of AppModel and all calls to GetIt to Settings.

Also included is the removal of diagnostic print statements.