Semi-successful implication of the getitmixin watch(...) method coupled to ValueNotifier.
SettingsServiceSeven extends ValueNotifier is registered within GetIt. The GetItMixin watch(...) method is used to extract settings data from the SettingsServiceSeven instance in GetIt.

Unfortunately the ValueNotifier approach doesn't work -- failing to generate a rebuild -- due to it relying on using the == operator inherent within ValueNotifier.

Suggest replacing ValueNotifier with an extension to ValueListenable.