Add BasePageView.
Abstract out the process of 'double-building' a widget with, firstly, nothing but an empty Container, and then with actual page contents in order for widgets further down the widget tree to get available screen dimensions via the GlobalKeyExtension.

Note the available screen dimensions may be different to 'MediaQuery.of(context).size' due to the presence of an AppBar and/or BottomAppBar.