Solves Issue #41.
Convert ButtonArray.rect method to a static getter.

Conversion of ButtonArray.rect has to have no inputs. Thus, instead of a BuildContext, an instance of Settings is retrieved from GetIt so that ButtonArray.rect can be calculated mathematically, rather than via the use of globalPaintBounds.

ButtonArray.rect uses Settings.basePageViewRect in order to translate and rotate the calculated Rect variable to the correct location. Settings.basePageViewRect is null initially, but during the post frame callback this is updated using the globalPaintBounds method within BasePageView.